+++ Game/Object/Entity/AttSys/AttCondition.cs
@@ -1,14 +1,5 @@
 namespace Goguma.Game.Object.Entity.AttSys
 {
-  public enum Cond
-  {
-    PlayerHp,
-    MonsterHp,
-    PlayerMaxHp,
-    MonsterMaxHp,
-    PlayerHpPer,
-    MonsterHpPer
-  }
   public class AttCondition
   {
     public Player.Player player;
