+++ Game/Object/Entity/AttSys/Cond.cs
@@ -1,0 +1,12 @@
+namespace Goguma.Game.Object.Entity.AttSys
+{
+  public enum Cond
+  {
+    PlayerHp,
+    MonsterHp,
+    PlayerMaxHp,
+    MonsterMaxHp,
+    PlayerHpPer,
+    MonsterHpPer
+  }
+}
